// frontend/src/proto/driver.ts
export interface DriverProfile {
    /**
     * @generated from protobuf field: string driver_id = 1
     */
    driverId: string;
    /**
     * @generated from protobuf field: string user_id = 2
     */
    userId: string;
    /**
     * @generated from protobuf field: string vehicle_no = 3
     */
    vehicleNo: string;
    /**
     * @generated from protobuf field: int32 capacity = 4
     */
    capacity: number;
    /**
     * @generated from protobuf field: repeated lastmile.driver.RoutePlan routes = 5
     */
    routes: RoutePlan[];
}
/**
 * @generated from protobuf message lastmile.driver.RoutePlan
 */
export interface RoutePlan {
    /**
     * @generated from protobuf field: string route_id = 1
     */
    routeId: string;
    /**
     * @generated from protobuf field: string driver_id = 2
     */
    driverId: string;
    /**
     * @generated from protobuf field: repeated lastmile.driver.RouteStop stops = 3
     */
    stops: RouteStop[];
    /**
     * @generated from protobuf field: string final_area_id = 4
     */
    finalAreaId: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp created_at = 5
     */
    createdAt?: Timestamp;
}
/**
 * @generated from protobuf message lastmile.driver.RouteStop
 */
export interface RouteStop {
    /**
     * @generated from protobuf field: int32 sequence = 1
     */
    sequence: number;
    /**
     * @generated from protobuf field: string area_id = 2
     */
    areaId: string;
    /**
     * @generated from protobuf field: bool is_station = 3
     */
    isStation: boolean;
    /**
     * @generated from protobuf field: int32 arrival_offset_minutes = 4
     */
    arrivalOffsetMinutes: number;
}
// ...
class DriverProfile$Type extends MessageType<DriverProfile> {
    constructor() {
        super("lastmile.driver.DriverProfile", [
            { no: 1, name: "driver_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "user_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "vehicle_no", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "capacity", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 5, name: "routes", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => RoutePlan }
        ]);
    }
    create(value?: PartialMessage<DriverProfile>): DriverProfile {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.driverId = "";
        message.userId = "";
        message.vehicleNo = "";
        message.capacity = 0;
        message.routes = [];
        if (value !== undefined)
            reflectionMergePartial<DriverProfile>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: DriverProfile): DriverProfile {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string driver_id */ 1:
                    message.driverId = reader.string();
                    break;
                case /* string user_id */ 2:
                    message.userId = reader.string();
                    break;
                case /* string vehicle_no */ 3:
                    message.vehicleNo = reader.string();
                    break;
                case /* int32 capacity */ 4:
                    message.capacity = reader.int32();
                    break;
                case /* repeated lastmile.driver.RoutePlan routes */ 5:
                    message.routes.push(RoutePlan.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: DriverProfile, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string driver_id = 1; */
        if (message.driverId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.driverId);
        /* string user_id = 2; */
        if (message.userId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.userId);
        /* string vehicle_no = 3; */
        if (message.vehicleNo !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.vehicleNo);
        /* int32 capacity = 4; */
        if (message.capacity !== 0)
            writer.tag(4, WireType.Varint).int32(message.capacity);
        /* repeated lastmile.driver.RoutePlan routes = 5; */
        for (let i = 0; i < message.routes.length; i++)
            RoutePlan.internalBinaryWrite(message.routes[i], writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message lastmile.driver.DriverProfile
 */
export const DriverProfile = new DriverProfile$Type();
// @generated message type with reflection information, may provide speed optimized methods
class RoutePlan$Type extends MessageType<RoutePlan> {
    constructor() {
        super("lastmile.driver.RoutePlan", [
            { no: 1, name: "route_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "driver_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "stops", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => RouteStop },
            { no: 4, name: "final_area_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "created_at", kind: "message", T: () => Timestamp }
        ]);
    }
    create(value?: PartialMessage<RoutePlan>): RoutePlan {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.routeId = "";
        message.driverId = "";
        message.stops = [];
        message.finalAreaId = "";
        if (value !== undefined)
            reflectionMergePartial<RoutePlan>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RoutePlan): RoutePlan {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string route_id */ 1:
                    message.routeId = reader.string();
                    break;
                case /* string driver_id */ 2:
                    message.driverId = reader.string();
                    break;
                case /* repeated lastmile.driver.RouteStop stops */ 3:
                    message.stops.push(RouteStop.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* string final_area_id */ 4:
                    message.finalAreaId = reader.string();
                    break;
                case /* google.protobuf.Timestamp created_at */ 5:
                    message.createdAt = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.createdAt);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RoutePlan, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string route_id = 1; */
        if (message.routeId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.routeId);
        /* string driver_id = 2; */
        if (message.driverId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.driverId);
        /* repeated lastmile.driver.RouteStop stops = 3; */
        for (let i = 0; i < message.stops.length; i++)
            RouteStop.internalBinaryWrite(message.stops[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* string final_area_id = 4; */
        if (message.finalAreaId !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.finalAreaId);
        /* google.protobuf.Timestamp created_at = 5; */
        if (message.createdAt)
            Timestamp.internalBinaryWrite(message.createdAt, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message lastmile.driver.RoutePlan
 */
export const RoutePlan = new RoutePlan$Type();
// ...
export const RouteStop = new RouteStop$Type();