// frontend/src/proto/google/protobuf/timestamp.ts
export interface Timestamp {
    /**
     * @generated from protobuf field: int64 seconds = 1
     */
    seconds: bigint;
    /**
     * @generated from protobuf field: int32 nanos = 2
     */
    nanos: number;
}
// ...
export const Timestamp = new Timestamp$Type();